Update README.md
# Design-of-an-active-filter-circuit-to-process-a-noisy-analog-signal
The task is to process a noisy signal. The only information given is the frequency of the wanted signal, it is ftarget = 42 kHz. My task is to design a suitable filter circuit to extract this signal, a SNR of 30
dB is required. There are several options to realize the requested filter circuit using active filter. For this assignment a PSpice project is provided by the Professor. The provided
project contains a voltage source generating the noisy signal. 
The detailed explanation of the given the task can be found on the file [Task To Do Analog Signal Assignment.pdf](https://github.com/MdSaifulIslamSajol/Design-of-an-active-filter-circuit-to-process-a-noisy-analog-signal/blob/8e5e8c70b2f9cd3e1a313a42e119e5708bb2cad9/Task%20To%20Do%20Analog%20Signal%20Assignment.pdf)